refactor: extract shared useDefaultRegionFilters hook
Deduplicate the preferred-regions logic from games-page and
console-games-page into a reusable hook.
