+++ web/src/pages/console-games-page.tsx
@@ -9,7 +9,6 @@ import {
   GameCardSkeleton,
   GameListRowSkeleton,
   EmptyState,
-  useToast,
 } from "@/components/ui";
 import { useConsoles } from "@/hooks/use-consoles";
 import { useGames, useToggleFavorite } from "@/hooks/use-games";
@@ -34,7 +33,7 @@ import { AlphabetBar } from "@/components/alphabet-bar";
 import { Pagination } from "@/components/pagination";
 import { BiosWarningBanner } from "@/features/bios/components/bios-warning-banner";
 import { useAuth } from "@/hooks/use-auth";
-import { useUserPreferences, useUpdatePreferences } from "@/hooks/use-preferences";
+import { useDefaultRegionFilters } from "@/hooks/use-default-region-filters";
 import type { GameFilters } from "@/types/api";
 
 type ViewMode = "grid" | "list";
@@ -111,10 +110,6 @@ export function ConsoleGamesPage() {
   const { toggle: handleTogglePlayLater } = useTogglePlayLater();
   const { data: biosData } = useBiosStatus();
   const { isAdmin } = useAuth();
-  const { data: userPrefs } = useUserPreferences();
-  const updatePrefs = useUpdatePreferences();
-  const { toast } = useToast();
-
   const console = consoles?.find((c) => c.id === id);
   const consoleName = console?.name ?? "Console";
 
@@ -126,18 +121,7 @@ export function ConsoleGamesPage() {
     ...parseUrlFilters(searchParams),
   }));
 
-  // Apply user's preferred regions as default when no regions are in the URL
-  const [appliedPreferredRegions, setAppliedPreferredRegions] = useState(false);
-  useEffect(() => {
-    if (
-      userPrefs?.preferredRegions?.length &&
-      !appliedPreferredRegions &&
-      !searchParams.get("regions")
-    ) {
-      setFilters((f) => ({ ...f, regions: userPrefs.preferredRegions }));
-      setAppliedPreferredRegions(true);
-    }
-  }, [userPrefs, appliedPreferredRegions, searchParams]);
+  const { saveDefaultRegions } = useDefaultRegionFilters(searchParams, setFilters);
 
   // Sync filters to URL (excluding consoleId)
   useEffect(() => {
@@ -229,12 +213,7 @@ export function ConsoleGamesPage() {
           isOpen={filtersOpen}
           onToggle={() => setFiltersOpen((o) => !o)}
           hideConsoleFilter
-          onSaveDefaultRegions={(regions) => {
-            updatePrefs.mutate(
-              { preferredRegions: regions },
-              { onSuccess: () => toast("info", "Default regions saved") },
-            );
-          }}
+          onSaveDefaultRegions={saveDefaultRegions}
         />
         <BestVersionsButton
           filters={filters}
